Rename "Rendering"-module to "Fluid"
