+++ Resources/Private/Templates/Fluid/General.html
@@ -1,6 +1,6 @@
 <html xmlns:f="http://typo3.org/ns/TYPO3/CMS/Fluid/ViewHelpers" data-namespace-typo3-fluid="true">
 <f:if condition="{fluidNamespaces}">
-    <h2 class="typo3-adminPanel-headline">Fluid global namespaces</h2>
+    <h2 class="typo3-adminPanel-headline">Global namespaces</h2>
 
     <table class="typo3-adminPanel-table typo3-adminPanel-table-debug">
         <thead>
@@ -32,7 +32,7 @@
 </f:if>
 
 <f:if condition="{preProcessors}">
-    <h2 class="typo3-adminPanel-headline">Fluid pre processors</h2>
+    <h2 class="typo3-adminPanel-headline">Pre processors</h2>
     <div class="typo3-adminPanel-table-overflow">
         <table class="typo3-adminPanel-table typo3-adminPanel-table-debug">
             <thead>
