+++ ext_localconf.php
@@ -33,15 +33,15 @@ call_user_func(
             );
         }
 
-        $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['adminpanel']['modules']['psychomieze_rendering'] = [
-            'module' => \Psychomieze\AdminpanelExtended\Modules\Rendering::class,
+        $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['adminpanel']['modules']['psychomieze_fluid'] = [
+            'module' => \Psychomieze\AdminpanelExtended\Modules\Fluid::class,
             'after' => ['debug'],
             'submodules' => [
                 'general' => [
-                    'module' => \Psychomieze\AdminpanelExtended\Modules\Rendering\General::class
+                    'module' => \Psychomieze\AdminpanelExtended\Modules\Fluid\General::class
                 ],
                 'fluid' => [
-                    'module' => \Psychomieze\AdminpanelExtended\Modules\Rendering\Fluid::class,
+                    'module' => \Psychomieze\AdminpanelExtended\Modules\Fluid\Fluid::class,
                     'after' => ['general']
                 ]
             ]
